Ajout du dossier frontend

+++ backend/controllers/user.js
@@ -26,7 +26,7 @@ exports.signUp = (req, res, next) => {
   const email = req.body.email;
   const pseudo = req.body.pseudo;
   const sql = `SELECT email FROM users WHERE email=?`;
-  //const sqlPseudo = `SELECT pseudo FROM users WHERE pseudo=?`;
+  const sqlPseudo = `SELECT pseudo FROM users WHERE pseudo=?`;
   let query = db.query(sql, email, async (err, docs) => {
     if (err) throw err;
 
@@ -36,14 +36,14 @@ exports.signUp = (req, res, next) => {
     if (docs.length === 1) {
       return res.status(400).json({ error: "Utilisateur déja existant !" });
     }
-    /*
+    
     let query = db.query(sqlPseudo, pseudo, async (err, docs) => {
       if (err) throw err;
       if (docs.length === 1) {
         return res.status(400).json({ error: "Pseudo déja existant !" });
       }
     });
-    */
+    
     if (password.length < 6) {
       return res.status(400).json({
         message: "Le mot de passe doit être de 6 caractéres minimum!",
@@ -71,6 +71,7 @@ exports.signUp = (req, res, next) => {
   });
 };
 
+// Passer le pseudo à l'utilisateur connecté ??
 // Connexion à un compte utilisateur
 exports.login = (req, res, next) => {
   const email = req.body.email;
@@ -91,12 +92,13 @@ exports.login = (req, res, next) => {
             .status(401)
             .json({ error: "Mot de passe incorrect" });
         }
-        const token = creatToken(docs[0].id);
+        //const token = creatToken(docs[0].id);
         // À vérifier ??
-        res.cookie('jwt', token, { httpOnly: true, maxAge });
+        //res.cookie('jwt', token, { httpOnly: true, maxAge });
         res.status(200).json({
           userId: docs[0].id,
-          token: jwt.sign({ userId: docs[0].id }, process.env.SECRET_TOKEN, {
+          //pseudo: docs[0].pseudo,
+          token: jwt.sign({ userId: docs[0].id/*, pseudo: docs[0].pseudo*/ }, process.env.SECRET_TOKEN, {
             expiresIn: "24h",
           }),
         });
@@ -117,7 +119,7 @@ exports.login = (req, res, next) => {
 // À corriger !!!!!
 // Déconnexion d'un compte utilisateur
 exports.logout = (req, res, next) => {
-  res.cookie('jwt', '', { maxAge : 1 }).json({message: "Utilisateur deconnecté !"});
+  //res.cookie('jwt', '', { maxAge : 1 }).json({message: "Utilisateur deconnecté !"});
   console.log("utilisateur connecté");
   //res.redirect('/');
 }
@@ -137,7 +139,7 @@ res.status(200).json({
 
 // Récupération des infos de tous les utilisateurs
 exports.getAllUsers = (req, res, next) => {
-  const sql = `SELECT email, pseudo, isAdmin, bio, picture FROM users`;
+  const sql = `SELECT id, email, pseudo, isAdmin, bio, picture FROM users`;
   let query = db.query(sql, (err, docs) => {
     if (err) throw err;
     res.status(200).json(docs);
@@ -147,12 +149,25 @@ exports.getAllUsers = (req, res, next) => {
 // Récupération des infos d'un utilisateur
 exports.userInfo = (req, res, next) => {
   const userId = req.params.id;
-  const sql = `SELECT email, pseudo, isAdmin, bio, picture FROM users WHERE id='${userId}'`;
+  const sql = `SELECT id, email, pseudo, isAdmin, bio, picture FROM users WHERE id='${userId}'`;[]
   let query = db.query(sql, (err, docs) => {
     if (err) throw err;
     res.status(200).json(docs);
   })
 };
+
+// À corriger !!!!!
+// Récupération des infos de l'utilisateur connecté
+exports.userProfil = (req, res, next) => {
+  const userId = req.auth.userId;
+  const sql = `SELECT id, email, pseudo, isAdmin, bio, picture FROM users WHERE id='${userId}'`;
+  let query = db.query(sql, (err, docs) => {
+    res.status(200).json(docs);
+    console.log(userId)
+    console.log(id)
+  })
+};
+
 
 // Quand l'utilisateur existe deja ? sinon enlever la possibilité de changer le pseudo !
 // Regler probleme de nodemon crash quand il y a des erreurs
@@ -194,6 +209,7 @@ exports.updateUser = (req, res, next) => {
               bio: bio,
               picture: new_profil_image_url,
             };
+            
             const sql = `UPDATE users SET ? WHERE id='${userPageId}'`;
             let query = db.query(sql, newUserInfos, (err, docs) => {
               if (err) {
@@ -254,33 +270,30 @@ exports.updateUser = (req, res, next) => {
 exports.deleteUser = (req, res, next) => {
   const userPageId = req.params.id;
   const userId = req.auth.userId;
-  const admin = req.body.admin;
   const sqlInfos = `SELECT id ,picture FROM users WHERE id='${userPageId}'`;
   const sqlAdminInfos = `SELECT isAdmin FROM users WHERE id='${userId}'`;
   let adminCheckout = null;
 
   let query = db.query(sqlAdminInfos, (err, docs) => {
     if (err) throw err;
-    adminCheckout = docs[0].admin;
+    adminCheckout = docs[0].isAdmin;
     let query = db.query(sqlInfos, (err, docs1) => {
       if (err) throw err;
       if (userId === docs1[0].id || adminCheckout === 1) {
-        const oldFileName = docs1[0].picture.split("/images/profils/")[1];
-
-        if (oldFileName !== "avatar.png") {
-          fs.unlink(`images/profils/${oldFileName}`, () => {
-            if (err) console.log(err);
-            else {
-              console.log("Image de profile supprimée");
-            }
-          });
+        if(docs1[0].picture) {
+          const oldFileName = docs1[0].picture.split("/images/profils/")[1];
+  
+          if (oldFileName !== "avatar.png") {
+            fs.unlink(`images/profils/${oldFileName}`, () => {
+              if (err) console.log(err);
+              else {
+                console.log("Image de profile supprimée");
+              }
+            });
+          }
         }
 
-        //const sql = `DELETE u.*,p.*, c.*,l.* FROM users u LEFT JOIN posts p ON(p.user_id = u.id) LEFT JOIN comments c ON(c.user_id = u.id) LEFT JOIN likes l ON (l.lp_utilisateur_id = u.id) WHERE u.id = '${userPageId}';`;
         const sql = `DELETE FROM users WHERE id = '${userPageId}';`;
-        const sqlPost = `DELETE * FROM posts WHERE posterId = '${userPageId}';`;
-        const sqlLike = `DELETE * FROM likes WHERE user_id = '${userPageId}';`;
-        const sqlComment = `DELETE * FROM comments WHERE user_id = '${userPageId}';`;
 
         
         let query = db.query(sql, (err, docs) => {
@@ -288,118 +301,11 @@ exports.deleteUser = (req, res, next) => {
           res.status(200).json({ message: "Utilisateur supprimé!" });
           console.log("utilisateur supprimé");
         });
-        
-        /*
-        let query = db.query(sql, (err, docs) => {
-          if (err) throw err;
-          res.status(200).json({ message: "Utilisateur supprimé !" });
-          console.log("utilisateur supprimé");
-          let query = db.query(sqlPost, (err, docs) => {
-            if (err) throw err;
-            res.status(200).json({ message: "Posts supprimés !" });
-            console.log("posts supprimé");
-            let query = db.query(sqlLike, (err, docs) => {
-              if (err) throw err;
-              res.status(200).json({ message: "Likes supprimés !" });
-              console.log("likes supprimé");
-              let query = db.query(sqlComment, (err, docs) => {
-                if (err) throw err;
-                res.status(200).json({ message: "Commentaires supprimés !" });
-                console.log("commentaires supprimé");
-              });
-            });
-          });
-        });
-        */
 
       } else {
         return res.status(403).json({ error: "Accès refusé" });
       }
     });
   });
-  console.log(`userId: ${userId} admin: ${admin} userPageId: ${userPageId}`);
+  console.log(`userId: ${userId} userPageId: ${userPageId}`);
 };
-
-
-
-
-
-/*
-// Abonnement à un utilisateur
-exports.follow = (req, res, next) => {
-  if (!objectID.isValid(req.params.id) || !objectID.isValid(req.body.idToFollow)) {
-    return res.status(400).send('ID unknown : ' + req.params.id)
-  }
-
-  try {
-    // add to the follower list
-    User.findByIdAndUpdate(
-      req.params.id,
-      { $addToSet: { following: req.body.idToFollow } },
-      { new: true, upsert: true },
-      (err, docs) => {
-        if (!err) {
-          res.status(201).json(docs);
-        }
-        else {
-          return res.status(400).json(err);
-        }
-      }
-    );
-    // add to following list
-    User.findByIdAndUpdate(
-      req.body.idToFollow,
-      { $addToSet: { followers: req.params.id } },
-      { new: true, upsert: true },
-      (err, docs) => {
-        //if (!err) {res.status(201).json(docs);}
-        if (err) {
-          return res.status(400).json(err);
-        }
-      }
-    );
-  }
-  catch (err) {
-    return res.status(500).json({ message: err });
-  }
-};
-
-// Désabonnement à un utilisateur
-exports.unfollow = (req, res, next) => {
-  if (!objectID.isValid(req.params.id) || !objectID.isValid(req.body.idToUnfollow)) {
-    return res.status(400).send('ID unknown : ' + req.params.id)
-  }
-
-  try {
-    // remove to the follower list
-    User.findByIdAndUpdate(
-      req.params.id,
-      { $pull: { following: req.body.idToUnfollow } },
-      { new: true, upsert: true },
-      (err, docs) => {
-        if (!err) {
-          res.status(201).json(docs);
-        }
-        else {
-          return res.status(400).json(err);
-        }
-      }
-    );
-    // remove to following list
-    User.findByIdAndUpdate(
-      req.body.idToUnfollow,
-      { $pull: { followers: req.params.id } },
-      { new: true, upsert: true },
-      (err, docs) => {
-        //if (!err) {res.status(201).json(docs);}
-        if (err) {
-          return res.status(400).json(err);
-        }
-      }
-    );
-  }
-  catch (err) {
-    return res.status(500).json({ message: err });
-  }
-};
-*/
